cognitive_backend/requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==0.28.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
gunicorn==21.2.0
